backend/pyproject.toml
[project]
name = "cheapbob"
version = "0.1.0"
description = "CheapBob Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.2.0",
    "alembic>=1.13.0",
    "apscheduler>=3.10.0",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.29.0",
    "beanie==1.26.0",
    "boto3==1.35.0",
    "fastapi[all]>=0.104.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "motor==3.3.2",
    "passlib[bcrypt]>=1.7.4",
    "portone-server-sdk==0.16.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymongo==4.6.3",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-socketio>=5.11.0",
    "pytz==2024.1",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]
